Introduction of Referer() and all its needed class
Introduction of socket into dependencies + Fix pylint + Introduction of Lookup() which can be called to execute `nslookup` or `whois` + Introduction of Percentage() which calculate and return the percentage of each status if needed + Introduction of generate which Generate a file in the desired location. + Introduction of Status() which return the domain status when whois is not satisfied or deactivated (future feature) + Introduction of Referer() which get the whois server we need to call.