MASSIVE REVIEW
This patch include a big changement in the orientation of PyFunceble.
Indeed, before this patch we were usint `tool.py` to do everything

that is needed to make `PyFunceble.py` works.

This patch includes a review of almost every lines.
This patch includes the usage of a more personalisable `config.yaml`
which allow us to changes for example the output directory with the
names we want.
This patch includes the future deletion of `tool.py` as every
functionalities of `tool.py` has been merged into `PyFunceble.py`