Improvement of the way to check for `INVALID` domains.
This patch is the result of the discussion at https://github.com/StevenBlack/hosts/issues/674.

This patch allow the testing of domain with `_` along with others (please report to the tests which are delivered with this path in order to learn about the different cases)

Thanks to @mitchellkrogza for providing some significant tests results and platform for the resolution of DNS, nslookup, dig and web-browsers while working with weird domain or hosts names with `_`.