Introduction of a way to avoid the retesting of already tested domains when  is set to .
Indeed, before this patch, we were testing over and over even if the domain was already present into our database. This patch correct that by deleting all domains/url which are already into the database. I mean by , the fact that a domain is not into the list which have to be tested in the current session (). I'm talking about all domains/url which are indexed with a timestamp.

Also:
  * Review of the way we handle the file content.'