Review of the way we remove already tested element from the list to test.
Indeed, we already did that before this patch. But because it was taking too long for very large database(s), a review was needed.

Please consider this patch as an improvement of the way we remove the element (from `inactive_db.json`) that we should not test before x day(s).

Thanks to @mitchellkrogza for reporting the issue.