Fix #23.
Indeed, before this patch, the autocontinue subsystem may not have worked proprely.

In fact, we were first cleaning and removing the `INACTIVE` element we
already tested before reading from the index of the last element we
tested.
This patch now get the part we need to test from the index of the last
element we tested and remove the element we already tested.

This method allow the size of the list to not change when we continue
from a previous session.

Contributors:
  * @dnmTX