Introduction of a fix of the multiprocessing logic.
Indeed, before this patch, a "block time" was possible.
It was not actually a "block time" but rather it looked like a block time as
we do not return anything.
It was actually the merge of all processes which was done.

Indeed, before this patch, I was waiting for the really end of the session
(autosave) of the file test (everything else) to merge the data which was
saved by all created processes. So if we had 15 000 processes,
we had 15 000 * (number of subsystem which saved data into that process)
index to merge into the different subsystems like for example
the autocontinue subsystem.

From now, we open x (given) simultaneous processes and at the end of all
of them I process the merge
of x * (number of subsystem which saved data into that process) index to merge.
This has the advantage that it fixes the previously mentionned issue but also
make the autocontinue subsystem more accurate and almost real-time.
The disavantage of this is we loose a bit (~3 seconds depending of the number
of processes to merge) between each creation of x (given) simultaneous
processes but compared to a test withouth multiprocessing it's nothing.

A big thanks to @mitchellkrogza && @dnmTX for spotting that 'blocking' issue.