Fix and improvement of along and around the multiprocessing logic.
1. Fix of a cleaning issue while using `--multiprocess`.

Indeed before this patch, there was a cleaning is issue
under `--multiprocessing` which caused the files to
sometimes be completely deleted in a middle of a test.

It was because I did not distingue the parent process
from its children.

Contributors:
    * @mitchellkrogza

2. Fix of the percentage counter

Indeed, before this patch, it was possible to get `300%` for example.

Contributors:
    * @mitchellkrogza

3. Reduction of function or code block call.

Indeed, under the autocontinue subsystem there was some codeblock which
were only needed under `--multiprocess` or at specific events under it.

4. Introduction of a `--multiprocess` configuration presets.

Indeed, before this patch, it was possible to give `0` as number of process.
It was also possible to disable the autocontinue subsystem even if it is
vital. It is now impossible to do that.

5. Fix issue when a link is given to `-f` or `-uf`.

Indeed, before this patch we were parsing the given URL as filename.
That caused the autocontinue subsystem to simply clean the whole output
directory.

6. Rename method for clarity.

Indeed, I renamed
    PyFunceble.file_multiprocess_core.FileMultiprocessCore.__auto_save_process
to
    PyFunceble.file_multiprocess_core.FileMultiprocessCore.__merge_processes_data