Introduction of a possible fix of the blocking issue under multiprocess.
Indeed, the previous patched related to that issue was
not working for huge lists.

This patch is a hope to fix that issue definitly.

Indeed, before this patch, I was ready each line and compare them with the
autocontinue database.
Now, I do that filter before each test which should kill that blocking issue.