Ensure that file are not generated in some special cases.
Indeed, before this patch, subjects which were present into the inactive
database were introduced into the outputted files on or after retest.

This patch blocks the generation of output for subjects which
are retested and the status is the same as the previous one.

Which mean for example if the previous status was `INVALID`,
and the new determined one is `INVALID` the tested subjects
will not be written in any of the `output/` destinations.

Contributors:
    * @dnmTX (@ultimate-hosts-blacklist)
    * @kowith337 (@dead-hosts)