Introduction of the `travis_distribution_branch` logic.
This patch fix the idea behind:
  * https://github.com/funilrys/PyFunceble/issues/44#issuecomment-537186271

Indeed, before this patch, we were systematically working from 1 branch
which is okay is most case. In some few cases you may want to have 2
branches. One for distribution of the results of PyFunceble and another
one for the processing of your list though PyFunceble.

This patch allows you to do that. In fact, you now use `--travis-branch`
to set the processing branch and `--travis-distribution-branch` to set
the processing branch.

Warning, we do not check if the branch exists. It should exists before
you even start PyFunceble or change their default values.

Contributors:
  * @mitchellkrogza
  * @spirillen