Reverse the way the installation with pip was documented.
Indeed, `python -m pip install pip --upgrade` is understandable because
we are updating pip which may be critical. Otherwise, I won't use it.

I'm willing to change this as soon as they change the command to
copy-paste from the official website.

As example, on top of our official PyPi repository:
  * https://pypi.org/project/PyFunceble-dev/

Also:
  * I don't know where does Python 2 come from but this project don't
    support Python 2 and will never do.