Improvement of the autocontinue behavior.
Indeed, before this patch, we were reading each line and checking if the
read lines should be ignored.

From now, before we start any test of a file, we create a shadow file
which will contain the lines that have not been (completely) tested.
This way, the system will continue almost directly.

The shadow file in question will be created in your temporary directory
and deleted once we don't need it. What it does is that it apply the
content of the given file against our method which checks if we should
ignore a given subject.

In other words, this patch let the testing subsystem focus on some real
test instead of waiting for new data.

This patch fixes https://git.io/JfOtZ

Possible side effects of this patch:
  You may wait a bit longer (depends on the size of your list) before
  the table is being printed. But - unless you have a huge list - this
  is not a big deal.

  In my test environment, the method which generates the shadow file
  took almost 10 seconds for ~ 9800 (hosts file) lines.

History:
  In the past, a similar mechanism existed but instead of writing into a
  shadow file, we had everything in memory which was not great.
  That's why I switched to the current process: one by one or multiple
  at one time for a multiprocessing use case.

Side Note:
  * If the `--dots` argument (or it's configuration equivalent) is
    given, you will see a brunch of `S` or `I` during the process of
    generating the shadow file.

    `S` stands for: Waiting to be tested (written in shadow file).
    `I` stands for: Ignore.

Also:
  * Cleanup of the `cleanup` method.

Contributors:
  * @mitchellkrogza
  * @spirillen