Prepare the switch from RAW SQL to SQLAlchemy.
Indeed, before this patch, I had to maintain RAW SQL for MySQL and MariaDB.
From now, it's all done through SQLALchemy and alembic for schema
migrations.

As of this commit, the migration (if available) will be executed
automatically.
As of this commit, the system won't try to download the MySQL and MariaDB
RAW SQL files anymore.

This patch includes:
  * Review / small improvement of the logic behind a MariaDB/MySQL
    database.
  * Bug fixes: Auto continue with multiprocessing and MariaDB/MysQL.