Remove the single producer thread.
This patch fixes #154.

Indeed, before this patch the STDOUT response may have looked "slow".
But in reality, it is because of the file writting which take a bit more
time than the STDOUT. Threfore, I created 2 new threads manager:

- 1 for the stdout.
- 2 for the dataset (output).

Writting to files, is rather slow and it is not new. But this patch
given end-user the impression that the result is already generated where
in reality it is not because we are trying to write files safely one line
at at time.