Fix issues regarding cleanup and end actions.
Indeed, before this patch, the way we were ending a test was misleading
because of the way I decided to handle the continuous integration.
This patch fixes the issue by create a condition and clear statement
about how things get cleaned-up when running under a CI engine or
normally (non CI-engine).