Introduction of the `--preload` argument and logic.
Indeed, before this patch, the resume from a previous session with the
autocontinue subsystem was taking a huge amount of time.
As of now, if the `--preload` argument - or its option counterpart -
is given, we decode and load the given input files into the continue
dataset before starting the test.
The preloading may take some time depending of the size of the file to
test, but this is the price for a smooth and better autocontinue.
Especially under CI. In fact, under CI, before this patch, some instances
were running only on auto continue comparison.

With this patch, instance under CI engines, will take the time they need
to preload the file to test and then process the continuous testing
without having to compare every single entries of the file with the
autocontinue dataset.

This patch touches #188, #193 and all issues related to the autocontinue
subsystem.