Introduction of the deletion of .PyFunceble_production.yaml.
This patch fixes #207.

Indeed, before this patch, end-user had the impression that it isn't
updated; we actually don't update nor touch it anymore because we don't
need it anymore.

Since PyFunceble 4.0.0, we embed the .PyFunceble_production.yaml
inside the distributed package of PyFunceble(-dev). Meaning that
as on now, we just read the distributed file which is located
inside the `data/infrastructure` directory of the PyFunceble
package.

Contributors:
  * @spirillen