Fix issue with preloading.
Indeed, before this patch, there was a bug which was preventing the
preloading of multiple files.
In other words, when multiple files are given with the --preload argument
the preloading of the second file stopped abruptly because I was trying
to start our core processes. But, they already existed at that time.

This patch fixes the issue by first checking if the process are
running.

Contributors:
  * @spirillen