Reduction of the usage of absolute path.
This patch fixes #215.

Indeed, before this patch, we were exclusively using absolute paths.
From now on, we will work as much as possible with relative paths.

This patch also includes the deletion of the 'source' column of the
inactive dataset along with some minor review of the fields to use for
comparison.

Contributors:
  * @spirillen