Improvement of the DNS Lookup.
This patch touches #221.

We now offer a way to bypass the default behavior of PyFunceble which
makes DNS lookup a bit slower. Indeed, the DNS Lookup is slow because it is not
only about the NS lookup. When PyFunceble test a subject over the DNS protocol,
it try to finds a positive result for the NS, CNAME, DNAME, A, AAAA records
over all the provided domains or subject.

In other words, what PyFunceble is doing (default), assuming we are testing
`example.org` and have 3 DNS servers given.

- Ask DNS 1 about, XXX.
    - If negative or erroneous. Continue to DNS 2
- Ask DNS 2 about, XXX.
    - If negative or erroneous. Continue to DNS 3
- Ask DNS 3 about, XXX.
    - If negative or erroneous. Take it as inactive.

But at each step if one is positive, we stop the loop on continue to the next
one. This process is done for each record (NS, CNAME, etc.) to query.

Now to the changes taken into this patch:

- Code simplification at some places.
- Subdomains will now be tested through the NS, A, AAAA, CNAME and DNAME in that
order.
- Second level domain will now be tested through the NS, CNAME, DNAME, A, AAAA
in that order.
- Introduction of the `--follow-server-order` argument, configuration index and
logic which let end-user determine if they want the order to be randomized.
- Introduction of the `--trust-dns-server` argument, configuration index and
logic which let end-user determine if they want to follow the following
instead of the previously presented one:

    - Ask DNS 1 about, XXX.
        - If negative. Take it as negative.
        - If erroneous. Continue to DNS 2
    - Ask DNS 2 about, XXX.
        - If negative. Take it as negative.
        - If erroneous. Continue to DNS 3
    - Ask DNS 3 about, XXX.
        - If negative. Take it as negative.
        - If erroneous. Take it as negative.

- Documentation of the new arguments and configuration indexes.