Rewrite AdBlock / Filter List decoder.
This patch closes #227.
This patch fixes #13.

To quote @keczuppp (#13):

> [.. ] but it seems you extract way too much in this mode on your own
> and it might cause troubles...

Therefore, I decided to rewrite the decoder completely.

This patch introduces a real split between what is normally decoded and
what is decoded within the aggressive mode.

Within the "standard" mode, we only decode what is supposed to be
blocked.
On the other side, within the "aggressive" mode, we decode
everything provided by the "standard" mode, plus everything behind a
'domain=' option or an 'href=' directive - if effective.

Please report to the tests to understand the differences on a more
deeper level and keep in mind that this new "direction" will evolve
with the time.

Decoding AdBlock or Filter lists is not an easy job and I hope to get
much more feedback in the future. I didn't implement this because
I have a use for it. But rather because it was asked by someone and
I wanted to see if I was capable of implementing it.

Now it's fully part of PyFunceble and people using it shouldn't be
afraid to submit the "weird things" they find while using the decoder.

Contributors:
  * @dnmTX
  * @jawz101
  * @keczuppp
  * @kulfoon
  * @spirillen