Improvement of the way we sort files or folders.
This patch provide a new worker which will let us sort a single file and
improve the one we use to sort the content of an output directory.

The reason behind this patch, is the fact that the mechanism I used
before this new implementation was not following the hierarchical sorting.
That means, in other words, that the hierarchical sorting of the outputed
files was broken.