Improvement of a process workflow.
Indeed, before this patch, we had to feed the worker with a message and
only the original worker that sent the feeding message could kill the
worker.

From now on, workers will still get the stop message. But their
dependencies won't get a stop message until all workers of a process
manager are stopped. This will prevent the stop message to mistakenly
send.