Do not follow "Retry-After" on retry.
Indeed, before this patch, we were following the "Retry-After" header -
that is given by the requested server.
As it is something that is not under our control, I decided to just stop
following it.
Therefore, from now on, all HTTP status code query won't follow the
"Retry-After" header to define the number of seconds to wait between each
retries.