Force HTTP status code lookup.
Indeed, before this patch, the HTTP status code was not systematically
queried in the case that the end-user allows our set of SPECIAL rules.

Therefore, if an end-user allows our set of SPECIAL rules, we are going
to try to query the HTTP status code before handling the status
to our SPECIAL rules handler.