Improvement of the HTTP Status code query tool.
This patch fixes #101.

Before this patch, the implementation was not "responsive"
enough. Indeed, we systematically provided the status code of the
provided URL (over port 80). That was a good idea back then, but in a
world where redirection and SSL/TLS certificate are cheap
(aka Let's Encrypt and ACME compatible CAs), the behavior should be
changed. And that's what this patch does.

With the patch, the HTTP Status code query tool will react differently
when it is "allowed" to follow redirect or not.

Here is what we used to do:

- http://example.org (302) -> STOP ===> 302

Here is what we do now - assuming that the interface is allowed to
follow or accept the results of the redirection:

- http://example.org (302)
    -> https://example.org (200) ===> 200

- http://example.org (302)
    -> https://test.example.rog (200) ===> 302`

- http://example.org (302)
    -> https://test.example.org (301)
        -> https://example.org (200) ===> 302

If the interface is allowed to accept the results of the redirection -
AKA `allow_redirects = True`, it will systematically provide the status
code at the end of the redirection.

Contributors:
  * @spirillen