Fix issue when a CNAME redirects to same CNAME.
This patch fixes #243.

Indeed, before this patch, and since we introduced the CNAME follow-up,
when a CNAME redicted to the same CNAME or a previous known one in the
chain, we were looping over and over.

This patch fixes it by always checking if any of the given CNAME was
already met.

Contributors:
  * @spirillen