Fix issue when a URL is incorrectly parsed.
This patch fixes #269.

Indeed, while using urllib.parse.urlpase, the hostname may be a NoneType.
Therefore, the following exception has been raised by other subsystem that
try to lookup the IP of hte hostname.

    TypeError: <self.subject> should be <class 'str'>, <class 'NoneType'>
    given.

This patch fixes it by double checking if a NoneType is sent for DNS
resolution.

Contributors:
  * @T145