Introduction of a friendly user setup checker.
This patch fixes #274.

Indeed, before this patch, there was no was to send some message to the
user when some common setup mistakes were found.
This patch fixes it by adding a "message collector" that is later printed
the end user.

For now on, the configuration checker only checks against #274, but this
patch opens the door for other improvement in the future.

Contributors:
  * @ZeroDot1
  * @spirillen