Extract the registrar from the WHOIS record.
This patch fixes #272.

Indeed, before this patch, there was no way to get access to the registrar
from the WHOIS record. This patch tries to solve the issue by adding a
registrar extractor.

This patch also introduce the `--registrar` and `--max-registrar` arguments
that can be used to display the top N registrar that were extracted.
Using the `--registrar` argument also generate a `logs/registrar.txt` file
that can be used to list the all registrar found while testing the current
file.

Necessary Migration: -- All migration will be automatically triggered.
  * csv file: The whois.csv file needs to be migrated in order for it to
    have the `registrar` column.
  * mariadb: The whois_record tables need to be migrated in order for it
    to have the `registrar` column.

Contributors:
  * @ryanbr