Add a way to differs worker startup and creation.
Indeed, before this patch, as soon as something is put inside an input
queue of a worker, the worker will be automatically created.
But, at sometime, for example for worker that should only run at the end,
we do not want that behavior.

Therefore, this patch introduce a way to differ the creation and startup
of a process - when necessary.
This patch apply for example for the file sorter that is actually only
necessary at the end of a test.