Do DNS Lookup before querying URL.
This patch fixes #164.

Indeed, before this patch, we were trying to query URL even if
the domain didn't have an IP address.
This patch fixes the issue by doing a classic (NS, CNAME, A, AAA)
query before trying to query any URL.

Note: The DNS Lookup is the same as the one we do for a standard
domain/ip test.

Contributors:
  * @mitchellkrogza
  * @spirillen