Better handling of proxy usage.
This patch touches #132.
This patch fixes #293.

Indeed, before this patch a special case related to the test on
onions subject wasn't handled properly.

This patch also replace the socks5 with socks5h protocol within
the documentation. User that are testing onions and do not have
a DNS resolver/server that can handle them are advised to use the
socks5h protocol instead of socks5. That will ensure that the DNS
lookup is not handled locally.

Contributors:
  * @spirillen