Fix issue when no nameserver is configured
This patch touches #328.

Indeed, in some rare case, no system-wide resolver/nameservers
may not be configured; causing our tests to fail.
This patch fixes the issue by testing and handling the case that
no resolver is given.

From now on, if no nameserver is configured, the default one will
be set instead to:
  - 1.1.1.1
  - 8.8.8.8
  - 9.9.9.9

Contributors:
  * @Nilsonfsilva