Introduction of a better split.
Indeed, before this patch, it was had to differ IP from domains.
This patch fixes the issue by introduction a new `subject_kind`
status key which can be used when a split between domains and
ips is necessary.

This patch also introduces a better recognition of domains and
IP when testing from the CLI.