Bugfix: Handle sets.
This patch fixes #358.

Indeed, before this patch, we always assumed the the given status
codes will either be given as list or tuple.

As of #359, sets were introduced which may produce some
unstability. Therefore, this patch fixes the issue by ensuring
that set are acceptable inputs when trying to switch status
based on the status code of the tested subject.

Contributors:
  * @spirillen