Fix type issue.
This patch fixes #360.

Indeed, before this patch, a new typing mechanism has been added into the collection library.
This means that by the time, the data hits the API, the given data is not a JSON, but a
python formatted type. In this case datetime.datetime.

This patch ensure that we handle all possible cases.

Contributors:
  * @spirillen