Add a locking mechanism when accessing CSV files.
This patch touches #435.

Indeed, as discovered by @Yuki2718 and @DandelionSprout, our processes sometimes
trigger a PermissionError exception on Windows.

While testing with one of the latest builds of Windows 11 (Pro), I was able to
reproduce the issue sometimes and therefore analyze it in more detail. It turns out
that in some rare cases, 2 processes are simultaneously attempting to read the same
CSV file.

This patch attempts to solve the problem by introducing a shared lock which - for now
- only affect Windows-Users. When running on Windows, all processes will share a lock
that blocks when 2 processes try to access the same ressource at the same time.

Contributors:
  * @Yuki2718
  * @DandelionSprout